requirements.txt
alembic==1.3.2
appdirs==1.4.4
attrs==19.3.0
autoenv==1.0.0
black==19.10b0
blinker==1.4
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
coverage==5.2.1
cryptography==2.8
flake8==3.8.3
Flask==1.1.1
Flask-Cors==3.0.7
Flask-Mail==0.9.1
flask-marshmallow==0.10.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
Flask-SSO==0.4.0
Flask-WTF==0.14.2
gunicorn==20.0.4
idna==2.8
install==1.3.3
itsdangerous==1.1.0
Jinja2==2.10.3
jwt==0.6.1
Mako==1.1.0
MarkupSafe==1.1.1
marshmallow==3.3.0
marshmallow-sqlalchemy==0.21.0
mccabe==0.6.1
more-itertools==8.4.0
msal==1.1.0
packaging==20.4
pathspec==0.8.0
pluggy==0.13.1
psycopg2-binary==2.8.4
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.19
pyflakes==2.2.0
PyJWT==1.7.1
pyparsing==2.4.7
pytest==5.4.3
pytest-cov==2.10.0
pytest-flask-sqlalchemy==1.0.2
pytest-mock==3.2.0
python-dateutil==2.8.1
python-dotenv==0.10.3
python-editor==1.0.4
regex==2020.7.14
requests==2.22.0
six==1.13.0
SQLAlchemy==1.3.12
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.8
wcwidth==0.2.5
Werkzeug==0.16.0
WTForms==2.2.1
